feat: implement trpc for library meta
